+++ lib/logic/cubit/app_theme_cubit.dart
@@ -87,4 +87,8 @@ class AppThemeCubit extends Cubit<AppThemeState> {
       }
     }
   }
+
+  Color getInProgress(){
+    return state.themeClass.successColor;
+  }
 }