+++ lib/presentation/utils/app_texts.dart
@@ -8,9 +8,11 @@ class AppTexts {
       builder: (context, appThemeState) {
         return Text(
           textProperties.text,
+          maxLines: textProperties.maxLines??1,
           style: TextStyle(
               fontSize: 24,
-              color:  textProperties.textColor ?? appThemeState.themeClass.textColorPrimary
+              color:  textProperties.textColor ?? appThemeState.themeClass.textColorPrimary,
+            decoration: textProperties.textDecoration??TextDecoration.none,
           ),
         );
       },
@@ -22,9 +24,11 @@ class AppTexts {
       builder: (context, appThemeState) {
         return Text(
           textProperties.text,
+          maxLines: textProperties.maxLines??1,
           style: TextStyle(
               fontSize: 16,
-              color:  textProperties.textColor ?? appThemeState.themeClass.textColorPrimary
+              color:  textProperties.textColor ?? appThemeState.themeClass.textColorPrimary,
+            decoration: textProperties.textDecoration??TextDecoration.none,
           ),
         );
       },
@@ -36,9 +40,11 @@ class AppTexts {
       builder: (context, appThemeState) {
         return Text(
           textProperties.text,
+          maxLines: textProperties.maxLines??1,
           style: TextStyle(
               fontSize: 12,
-              color:  textProperties.textColor ?? appThemeState.themeClass.textColorPrimary
+              color:  textProperties.textColor ?? appThemeState.themeClass.textColorPrimary,
+            decoration: textProperties.textDecoration??TextDecoration.none,
           ),
         );
       },
@@ -50,9 +56,11 @@ class AppTexts {
       builder: (context, appThemeState) {
         return Text(
           textProperties.text,
+          maxLines: textProperties.maxLines??1,
           style: TextStyle(
               fontSize: 10,
-              color:  textProperties.textColor ?? appThemeState.themeClass.textCaptionColor
+              color:  textProperties.textColor ?? appThemeState.themeClass.textCaptionColor,
+            decoration: textProperties.textDecoration??TextDecoration.none,
           ),
         );
       },
@@ -63,9 +71,13 @@ class AppTexts {
 class TextProperties {
   String text;
   Color? textColor;
+  int? maxLines;
+  TextDecoration? textDecoration;
 
   TextProperties({
     required this.text,
     this.textColor,
+    this.maxLines,
+    this.textDecoration,
   });
 }